chore(git): add official code examples as a git module
This patch adds AWS lambda in action official code examples from the
remote repository as a git submodules as a reference to this repository.

note: when the book was released only node 4.x was available as a valid
runtime for AWS Lambda. As node 6.x is now a valid runtime all examples
are updated to supported ES6 semantic in my own version of the example.